Restrict fr-minecraft to distribution paths
Site owner has indicated by email that the purpose of the site is not to distribute mods, and that the redistribution of copyrighted mods on this site is a side effect of lack of moderation capacity. This does not excuse the site in any way. The rating has been updated to more accurately reflect the state of the site while still effectively warning visitors to the mod pages that the site hosts mods illegally.

Further, clarification about malicious software is made because of reports of competing sites using the StopModReposts policy and philosophy for wrongful defamation of the site.